+++ CHANGELOG.md
@@ -4,8 +4,9 @@ All notable changes to the "al-codeactions" extension will be documented in this
 
 ## 0.2.15
 
-- Fix implicit with-usages
+- Command added: Fix implicit with-usages
   - fixes all warnings displayed in the problems pane. If there are more than 100 warnings inside a file, the command has to be executed again at a later time.
+- Command added: Add pragma 'implicit with disable' to all files.
 
 ## 0.2.14